@@ -58,7 +58,7 @@ export function useGameState(gameMode: GameMode) {
 
     // Uses setInterval to run the gameLoop every x seconds. If the status
     // changes to "Paused", I think the cleanup will remove the setInterval from
-    // the previous run, and then retrn early, thus not setting up another set
+    // the previous run, and then return early, thus not setting up another set
     // interval and pausing the game.
     useEffect(() => {
         const gameLoop = () => {
